+++ src/pages/groupPage.tsx
@@ -61,30 +61,30 @@ const GroupPage = (): JSX.Element => {
     const selectedGroup = reduxState.group.group;
 
     return (
-        <div className="container body">
+        <React.Fragment>
             <SearchBar />
             {reduxState.thunk.fetch_in_progress ? (
-                <div className="center_container">
+                <div className="spinner">
                     <Spinner animation="border" />
                 </div>
             ) : (
                 <>
                     {reduxState.thunk.fetch_failed ? (
-                        <div>
+                        <React.Fragment>
                             <FetchError />
-                        </div>
+                        </React.Fragment>
                     ) : (
-                        <div>
+                        <React.Fragment>
                             {selectedGroup && (
                                 <GroupInfo title={selectedGroup.name} description={selectedGroup.description} />
                             )}
                             <h3>Våre lag</h3>
                             {listContent.length === 0 ? <EmptyResult /> : <div className={cardList}>{listContent}</div>}
-                        </div>
+                        </React.Fragment>
                     )}
                 </>
             )}
-        </div>
+        </React.Fragment>
     );
 };
 export default GroupPage;
